requirements.txt
streamlit==1.45.1
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.6.1
tensorflow_cpu==2.19.0
keras==3.10.0
streamlit-option-menu==0.4.0
folium==0.19.6
streamlit_folium==0.25.0